re-added a lot of const ,-) Thanks frerich for pointing out that my last check in made little sense. Here is what the problem really was:
I had one class TextRangeRegExp (inheriting from RegExp) that did not have
const in its operator==, and that had the consequence that the builtin
operator== was used when I compared to RegExp& with == ( when they really
were TextRangeRegExp). This of course had less to do with const issues than
with overriding methods in subclasses.

svn path=/trunk/kdeutils/kregexpeditor/; revision=232089